Aalto recognizer, Turku parser: from Testing to Speech, Parsing, resp.
Also rewrote their human-visible names somewhat. That's easy to do.

Removed old Aalto tools from the Mylly interface. The batch processing
pair of tools should be the only interface to the speech analysis
engines, because they are expected to be heavier than one likes to run
in taito-shell.

The parser is still also runnable in taito-shell.