Rethinking batch jobs: submit and wait, optionally check state
This replaces the previous mechanism. Testing the idea with the
parser, but already thinking that this is better in every way:
simpler, more flexible, easier to use and understand, usable to build
Chipster workflows (which the other one may not have been).

Initial tests outside Chipster seemed to go as intended.