Add first Korp API tool
Getting a concordance from Suomi24 (just the S24 sample for now, but
probably expand for more)

The plan is to have different tools for different corpora so that
the selection of standard corpora in Mylly is easy and the tool can
know what attributes are available in the particular corpus.

The output is JSON. The plan is to provide tools to transform it to
various forms.