A fuller version of TSV as ODS
Base column types on column name *prefix* - we attempt to do this
systematically from now on, rely on the user to rename attributes to
get the benefit of this, and have our own tools produce such names.

cM, kM indicate integer type (count and otherwise)
wM, vM indicate floating type (weight or otherwise)

That's M for Mylly, there might be further prefixes, and the slight
sense of weird is an attempt to avoid accidental conflicts. And we
want it to be short, right?