Make Korp query have random seed (default random) and page
User needs to repeat the query specifically to get the next page, to
not hammer on Korp unnecessarily, but when they want that, they can
get that, within limits. Korp apparently does not cache the search
results for mere fetching of further results, so it will repeat the
whole CQP search.