Attempt to detect and report error response from Korp
Previously error reports (also in JSON) have been simply dumped to
a file, same as actual results. Now detect (hopefully) and report
instead. Note: query is temporarily intentionally broken so as to
test that the error detection code works (which it probably not,
judging from the previous patchwork of fixes this morning, but
how is one to know until one tries it and sees - sigh, ok?)