Added access to two more corpora (COCA and COHA)
New corpora have not yet their own attributes! But Suomi24
attributes. Not sure if the tools will even work if those
are missing. Probably? But they will not produce attributes
that are missing in Suomi24 which needs fixed in a next step.

But there are plenty of corpora to incorporate as tools.
One made a list in the file corpora-list. How does one
even find out what attributes they have?