Decoder of Windows-1252
New plan is to have UTF-8 as implicit target, transcoding tools be
really simple to run with no parameters at all

There appears to be no way to prevent transcoding even such codes as
are not part of the source code, though? Not sure. On the other hand,
the real problem is control codes that are in some sense valid input.